requirements.txt
altgraph==0.17.4
beautifulsoup4==4.13.5
bidict==0.23.1
blinker==1.9.0
bs4==0.0.2
build==1.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
docutils==0.22
EbookLib @ git+https://github.com/aerkalov/ebooklib
filetype==1.2.0
Flask==3.1.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
h11==0.16.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
lxml==6.0.1
MarkupSafe==3.0.2
ml_dtypes==0.5.3
mpmath==1.3.0
nltk==3.9.1
numpy==2.3.2
onnxruntime==1.22.1
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
piper-tts==1.3.0
protobuf==6.32.0
pycparser==2.22
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pypdf==5.9.0
pypiwin32==223
pyproject_hooks==1.2.0
pypub==1.5
pyreadline3==3.5.4
pytesseract==0.3.13
python-engineio==4.12.2
python-socketio==5.13.0
pywin32==311
pywin32-ctypes==0.2.3
regex==2025.7.34
requests==2.32.5
setuptools==80.9.0
sherpa-onnx==1.12.11
sherpa-onnx-core==1.12.11
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.8
sympy==1.14.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
